v 2.0

+++ src/main/java/ru/hogwarts/school/model/AverageAge.java
@@ -1,5 +1,5 @@
 package ru.hogwarts.school.model;
 
 public interface AverageAge {
-    Integer getAverageAge();
+    Float getAverageAge();
 }
